pyproject.toml
[tool.poetry]
name = "insight-chat"
version = "0.1.0"
description = ""
authors = ["hieuGoku <ngochieu02.ai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
qdrant-client = "^1.7.1"
pypdf = "^4.0.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.1"
pymongo = "^4.6.1"
motor = "^3.3.2"
llama-index-vector-stores-qdrant = "^0.1.1"
llama-index-storage-docstore-mongodb = "^0.1.1"
llama-index-storage-index-store-mongodb = "^0.1.1"
llama-index = "^0.10.5"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.2"
fake-useragent = "^1.4.0"
llama-index-readers-remote = "^0.1.4"
llama-index-llms-openai = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
